+++ src/main/java/org/sopt/solply_server/domain/user/service/UserOnboardingService.java
@@ -29,60 +29,12 @@ import org.springframework.transaction.annotation.Transactional;
 public class UserOnboardingService {
 
     private final UserRepository userRepository;
-    private final TownService townService;
     private final UserValidator userValidator;
-    private final UserInterestTownService userInterestTownService;
 
-    @Transactional
-    public UserUpdateResponse completeOnboarding(Long userId, UserUpdateRequest request) {
-        // 사용자 조회
+    public UserPersonaListGetResponse getUserPersonaList(Long userId) {
         User user = userRepository.findById(userId)
                 .orElseThrow(() -> new BusinessException(ErrorCode.NOT_FOUND_USER));
-
-        // 비즈니스 규칙 검증
         userValidator.validateOnboardingAvailable(user);
-        Town town = townService.findTownById(request.favoriteTown());
-        userValidator.validateNicknameNotDuplicated(request.nickname());
-
-        // 온보딩 정보 업데이트 (재시도 로직 적용)
-        return updateUserWithRetry(user, town, request);
-    }
-
-    /**
-     * DataIntegrityViolationException에만 재시도 적용
-     * 동시성 처리 전략:
-     * 1. 애플리케이션 레벨 사전 검증 (빠른 실패)
-     * 2. DB UNIQUE 제약조건 (최종 방어선)
-     * 3. @Retryable을 통한 재시도 (일시적 실패 대응)
-     */
-    @Transactional
-    @Retryable(
-            retryFor = {DataIntegrityViolationException.class},
-            maxAttempts = 3,
-            backoff = @Backoff(delay = 100, multiplier = 2)
-    )
-    public UserUpdateResponse updateUserWithRetry(User user, Town town, UserUpdateRequest request) {
-        try {
-            // 온보딩 정보 업데이트
-            user.updateOnboardingInfo(request.persona(), request.nickname());
-            User savedUser = userRepository.save(user);
-
-            // 관심 동네 저장
-            userInterestTownService.saveUserInterestTown(savedUser, town);
-
-            log.info("온보딩 완료: userId={}, nickname={}, townId={}",
-                    user.getId(), request.nickname(), town.getId());
-
-            return UserUpdateResponse.of(savedUser, town);
-
-        } catch (DataIntegrityViolationException e) {
-            log.warn("DB 제약조건 위반으로 인한 온보딩 실패: userId={}, nickname={}",
-                    user.getId(), request.nickname());
-            throw e; // 재시도를 위해 예외를 다시 던짐
-        }
-    }
-
-    public UserPersonaListGetResponse getUserPersonaList() {
         List<UserPersonaDto> personaDtos = Arrays.stream(UserPersona.values())
                 .map(UserPersonaDto::from)
                 .toList();
